pocs/inner_lens/requirements.txt
streamlit>=1.30.0
anthropic
opencv-python
mediapipe
python-dotenv
